Add notifications controller
