+++ app.py
@@ -2,25 +2,19 @@ import os
 from flask_cors import cross_origin
 from api import app
 from api.util.decorators import token_required
-from api.services.notifications import HandleUserNotification
 from api.services.comments import Comentarios, ComentariosPostagem
 from api.services.posts import Filtros, Categorias, ListaPostagens, Postagens, PostagensId, Recomendados, Selo
-from api.controller import users, auth, forms
+from api.controller import users, auth, forms, notifications
 
 app.register_blueprint(users.app)
 app.register_blueprint(auth.app)
 app.register_blueprint(forms.app)
+app.register_blueprint(notifications.app)
 
 @app.route('/')
 @cross_origin(origin='*', headers=['Content-Type', 'Authorization'])
 def hello():
 	return "This API Works! [" + os.environ.get("ENV", "DEV") + "]"
-
-@app.route('/users/<id>/notificacoes_conf', methods=['PUT', 'GET'])
-@cross_origin(origin='*', headers=['Content-Type', 'Authorization'])
-@token_required
-def handle_user_notificacao(id):
-    return HandleUserNotification(id)
 
 @app.route('/categorias', methods=['POST', 'GET'])
 @cross_origin(origin='*', headers=['Content-Type', 'Authorization'])
